# specs/002-react-admin-frontend/contracts/openapi-admin.yaml
openapi: 3.0.3
info:
  title: Admin API Endpoints
  description: Backend API endpoints for administrative operations in multi-tenant system. Frontend-agnostic APIs supporting any client framework.
  version: 1.0.0
  contact:
    name: Backend API Team
  license:
    name: MIT
servers:
  - url: /api/v1
    description: API v1 base path (admin endpoints)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required: [user_id, tenant_id, email, is_disabled, failed_login_attempts, roles, created_at, updated_at, created_by, updated_by]
      properties:
        user_id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          description: Unique within tenant
        is_disabled:
          type: boolean
        failed_login_attempts:
          type: integer
          minimum: 0
        last_login_at:
          type: string
          format: date-time
          nullable: true
        full_name:
          type: string
          nullable: true
        job_title:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        timezone:
          type: string
          description: IANA timezone identifier
          nullable: true
        language:
          type: string
          description: ISO 639-1 language code
          nullable: true
        roles:
          type: array
          items:
            type: string
            enum: [superadmin, tenant_admin, developer, analyst, user, service_account, support_readonly]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: uuid
        updated_by:
          type: string
          format: uuid
...
    ListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items: {}
        pagination:
          type: object
          required: [page, per_page, total, total_pages]
          properties:
            page:
              type: integer
              minimum: 1
            per_page:
              type: integer
              minimum: 1
              maximum: 100
            total:
              type: integer
              minimum: 0
            total_pages:
              type: integer
              minimum: 0
    SingleResponse:
      type: object
      required: [data]
      properties:
        data: {}
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message]
          properties:
            message:
              type: string
            code:
              type: string
            details:
              type: object

    # Create/Update Request Schemas
...
    CreateUserRequest:
      type: object
      required: [email, roles]
      properties:
        email:
          type: string
          format: email
        full_name:
          type: string
        job_title:
          type: string
        department:
          type: string
        phone:
          type: string
        timezone:
          type: string
        language:
          type: string
        roles:
          type: array
          items:
            type: string
        is_disabled:
          type: boolean
          default: false
        password:
          type: string
          minLength: 8
          description: Optional - if not provided, invitation will be sent
    UpdateUserRequest:
      type: object
      description: |
        Update core user fields (email, roles, status).
        
        Note: Extended profile fields (full_name, job_title, department, phone, timezone, language)
        are managed in the user_details table (feature 003-user-profile-details) and should be
        updated via PUT /api/v1/users/{user_id}/profile endpoint.
      properties:
        email:
          type: string
          format: email
          description: User email address (must be unique within tenant)
        roles:
          type: array
          items:
            type: string
          description: User roles (superadmin assignment requires superadmin permission)
        is_disabled:
          type: boolean
          description: Account disabled status
...
  parameters:
    TenantIdQuery:
      name: tenant_id
      in: query
      description: Filter by tenant ID (required for superadmin cross-tenant operations)
      schema:
        type: string
        format: uuid
    PageQuery:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageQuery:
      name: per_page
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    SortQuery:
      name: sort
      in: query
      description: Sort field
      schema:
        type: string
        default: created_at
    OrderQuery:
      name: order
      in: query
      description: Sort direction
      schema:
        type: string
        enum: [asc, desc]
        default: desc
paths:
...
  /users:
    get:
      summary: List users
      description: List users within authorized scope (tenant-filtered for non-superadmin)
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PerPageQuery'
        - $ref: '#/components/parameters/SortQuery'
        - $ref: '#/components/parameters/OrderQuery'
        - name: role
          in: query
          description: Filter by role
          schema:
            type: string
        - name: is_disabled
          in: query
          description: Filter by status
          schema:
            type: boolean
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ListResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'

    post:
      summary: Create user
      description: Create a new user within authorized tenant scope
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
  /users/{user_id}:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      summary: Get user by ID
      description: Get specific user details within authorized scope
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'

    put:
      summary: Update user
      description: Update user details within authorized scope
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'

    delete:
      summary: Delete user
      description: Delete user within authorized scope (soft delete - sets status to disabled)
      responses:
        '204':
          description: User deleted
  /users/{user_id}/reset-password:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      summary: Reset user password (admin-initiated)
      description: |
        Admin-initiated password reset without requiring current password verification.
        
        Authorization:
        - Tenant admins can reset passwords for users in their tenant
        - Superadmins can reset passwords for any user
        - Only superadmins can reset superadmin passwords
        
        Note: This is different from self-service password change which requires current password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_password]
              properties:
                new_password:
                  type: string
                  minLength: 8
                  description: New password (must meet strength requirements)
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user_id:
                    type: string
                    format: uuid
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{user_id}/restore:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      summary: Restore disabled user
      description: Restore a soft-deleted (disabled) user account within authorized scope
      responses:
        '200':
          description: User restored
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                    format: uuid
                  status:
                    type: string
        '400':
          description: User is not disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Policy Management Endpoints
...